fix: inline job processors on TeamsBridgeApp to fix undefined app in scheduler
- Remove jobs/InboundNotificationProcessor.ts; inline all three job processors
  as arrow function properties on TeamsBridgeApp, closing over
- lib/InboundNotificationHelper.ts: import type TeamsBridgeApp (breaks CommonJS
  circular dep cycle that caused 'not a constructor' crash)
- batch-o: split MicrosoftGraphApi into lib/graph/ (28 function files + barrel)
- docs/refactor-progress.md: updated commit history and TeamsBridgeApp status
