+++ lib/graph/deleteSubscriptionAsync.ts
@@ -1,0 +1,24 @@
+import { HttpStatusCode, IHttp, IHttpRequest } from "@rocket.chat/apps-engine/definition/accessors";
+import { getGraphApiSubscriptionOperationUrl } from "../Const";
+
+export const deleteSubscriptionAsync = async (
+    http: IHttp,
+    subscriptionId: string,
+    userAccessToken: string): Promise<void> => {
+    const url = getGraphApiSubscriptionOperationUrl(subscriptionId);
+
+    const httpRequest: IHttpRequest = {
+        headers: {
+            'Content-Type': 'application/json',
+            'Authorization': `Bearer ${userAccessToken}`,
+        },
+    };
+
+    const response = await http.del(url, httpRequest);
+
+    if (response.statusCode === HttpStatusCode.NO_CONTENT) {
+        return;
+    } else {
+        throw new Error(`Delete subscription failed with http status code ${response.statusCode}.`);
+    }
+};
